Funcionalidad: validacion de parametros en los controladores

+++ src/payments/payments/payments.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, InternalServerErrorException } from '@nestjs/common';
 import { PaymentsService } from './payments.service';
 import { CreatePaymentDto } from './dto/create-payment.dto';
 import { UpdatePaymentDto } from './dto/update-payment.dto';
@@ -6,6 +6,7 @@ import { Roles } from 'src/auth/roles.decorator';
 import { UserRole } from 'src/users/entities/user/user.entity';
 import { plainToInstance } from 'class-transformer';
 import { PaymentResponseDto } from './dto/payment-response.dto';
+import { validateParameters } from 'src/shared/utils/parameters-validation';
 
 @Controller('payments')
 export class PaymentsController {
@@ -27,6 +28,7 @@ export class PaymentsController {
   @Get(':id')
   @Roles(UserRole.CLIENT, UserRole.ADMIN)
   async findOneDto(@Param('id') id: string) {
+    if(!validateParameters(id)) throw new InternalServerErrorException('Parametros inválidos')
     const payment = await this.paymentsService.findOne(+id);
     if (!payment) return null;
     return plainToInstance(PaymentResponseDto, payment, { excludeExtraneousValues: true });
@@ -35,18 +37,21 @@ export class PaymentsController {
   @Patch(':id')
   @Roles(UserRole.CLIENT, UserRole.ADMIN)
   update(@Param('id') id: string, @Body() updatePaymentDto: UpdatePaymentDto) {
+    if(!validateParameters(id)) throw new InternalServerErrorException('Parametros inválidos')
     return this.paymentsService.update(+id, updatePaymentDto);
   }
 
   @Delete(':id')
   @Roles(UserRole.ADMIN)
   remove(@Param('id') id: string) {
+    if(!validateParameters(id)) throw new InternalServerErrorException('Parametros inválidos')
     return this.paymentsService.delete(+id);
   }
 
   @Patch(':id/confirm')
   @Roles(UserRole.ADMIN) 
   async confirmPayment( @Param('id') id: number, @Body('success') success: boolean) {
+    if(!validateParameters(id)) throw new InternalServerErrorException('Parametros inválidos')
     return this.paymentsService.confirm(+id, success);
   }
 }
