+++ src/products/products.controller.ts
@@ -6,6 +6,7 @@ import { Roles } from 'src/auth/roles.decorator';
 import { UserRole } from 'src/users/entities/user/user.entity';
 import { Public } from 'src/auth/public.decorator';
 import { FilterProductDto } from './entities/dto/filter-product.dto';
+import { validateParameters } from 'src/shared/utils/parameters-validation';
 
 
 @Controller('products')
@@ -16,6 +17,7 @@ export class ProductsController {
 	@Public()
 	findAll(@Query('page') page?: string, @Query('limit') limit?: string, @Query('isActive') isActive?: boolean,
 	@Query('category') category?: string) {
+		if(!validateParameters(page, limit)) throw new InternalServerErrorException('Parametros inválidos')
 		const options: any = {};
 		if (page) options.page = Number(page);
 		if (limit) options.limit = Number(limit);
@@ -29,13 +31,13 @@ export class ProductsController {
 	@Get(':id')
 	@Public()
 	findOne(@Param('id') id: string) {
+		if(!validateParameters(id)) throw new InternalServerErrorException('Parametros inválidos')
 		return this.productsService.findOne(+id);
 	}
 
 	@Post()
 	@Roles(UserRole.VENDOR, UserRole.ADMIN)
 	create(@Body() createProductDto: CreateProductDto, @Request() req) {
-
 		createProductDto.vendorId = req.user.vendorProfileId
 		return this.productsService.create(createProductDto);
 	}
@@ -43,6 +45,7 @@ export class ProductsController {
 	@Patch(':id')
 	@Roles(UserRole.VENDOR, UserRole.ADMIN)
 	async update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto, @Request() req) {
+		if(!validateParameters(id)) throw new InternalServerErrorException('Parametros inválidos')
         const product = await this.productsService.findOne(+id)
 		console.log(product + "" + req.user.vendorProfileId + " " + product?.vendor.id)
 		if (!product || product.vendor.id !== req.user.vendorProfileId) throw new InternalServerErrorException('Producto no encontrado o registrado como propio')
@@ -52,6 +55,7 @@ export class ProductsController {
 	@Delete(':id')
 	@Roles(UserRole.VENDOR, UserRole.ADMIN)
 	async remove(@Param('id') id: string, @Request() req) {
+		if(!validateParameters(id)) throw new InternalServerErrorException('Parametros inválidos')
 		const product = await this.productsService.findOne(+id)
 		if (!product || product.vendor.id !== req.user.vendorProfileId) throw new InternalServerErrorException('Producto no encontrado o registrado como propio')
 		return this.productsService.delete(+id);
